requirements.txt
django
django-debug-toolbar
cookiecutter
django-extensions
django-crispy-forms
django-rest-swagger
django-filter
djoser
sphinx

